+++ middleware/azure_iot/c-utility/src/sastoken.c
@@ -23,7 +23,7 @@ static double getExpiryValue(const char* expiryASCII)
     {
         if (expiryASCII[i] >= '0' && expiryASCII[i] <= '9')
         {
-            value = value * 10 + (expiryASCII[i] - '0');
+            value = value * 10 + (double)(expiryASCII[i] - '0');
         }
         else
         {
@@ -165,7 +165,7 @@ bool SASToken_Validate(STRING_HANDLE sasToken)
                     memset(expiryASCII, 0, seStop - seStart + 1);
                     for (i = seStart; i < seStop; i++)
                     {
-                        // The se contains the expiration values, if a & token is encountered then 
+                        // The se contains the expiration values, if a & token is encountered then
                         // the se field is complete.
                         if (sasTokenArray[i] == '&')
                         {
